requirements.txt
python-telegram-bot[ext]==21.4
aiohttp==3.9.5
